DOMA-1722 extracted `SbbolCredentials` class to `integrations/sbbol` folder
